Initial commit of the tock-registers crate repository
This commit serves as the root of the split-out tock-registers crate
repository. Whereas tock-registers used to be maintained within the
Tock repository, coinciding with the first (1.0) major release it is
to be split out into its own, separate repository.

To give this repository a clean starting point the root is a new
commit, including the LICENSE terms which used to be located at the
Tock repository root. To honor the contributions by previous authors
and preserve authorship information, a filtered tree of the Tock
repository containing all relevant tock-registers commits will be
merged in.

Signed-off-by: Leon Schuermann <leon@is.currently.online>